# community/go-engines-community/docs/swagger.yaml
basePath: /api/v4
definitions:
...
  common.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  common.PaginatedListResponse:
    properties:
      data:
        type: object
      meta:
        $ref: '#/definitions/common.PaginatedMeta'
        type: object
    type: object
  common.PaginatedMeta:
    properties:
      page:
        type: integer
      page_count:
        type: integer
      per_page:
        type: integer
      total_count:
        type: integer
    type: object
  common.ValidationErrorResponse:
    properties:
      errors:
        additionalProperties:
          type: string
        type: object
    type: object
  contextgraph.ConfigurationItem:
    properties:
      _id:
        type: string
      action:
        type: string
      action_properties:
        type: object
      category:
        type: string
      enabled:
        type: boolean
      entity_patterns:
        $ref: '#/definitions/pattern.EntityPatternList'
        type: object
      impact_level:
        type: integer
      infos:
        additionalProperties: true
        type: object
      measurements:
        items:
          type: object
        type: array
      name:
        type: string
      output_template:
        type: string
      type:
        type: string
    type: object
  contextgraph.ImportJob:
    properties:
      creation:
        type: string
      execTime:
        type: string
      id:
        type: string
      info:
        type: string
      stats:
        $ref: '#/definitions/contextgraph.JobStats'
        type: object
      status:
        type: string
    type: object
  contextgraph.ImportResponse:
    properties:
      _id:
        type: string
    type: object
  contextgraph.JobStats:
    properties:
      deleted:
        type: integer
      updated:
        type: integer
    type: object
  contextgraph.Link:
    properties:
      _id:
        type: string
      action:
        type: string
      action_properties:
        type: object
      from:
        items:
          type: string
        type: array
      infos:
        additionalProperties: true
        type: object
      to:
        type: string
    type: object
  contextgraph.Request:
    properties:
      json:
        properties:
          cis:
            items:
              $ref: '#/definitions/contextgraph.ConfigurationItem'
            type: array
          links:
            items:
              $ref: '#/definitions/contextgraph.Link'
            type: array
        type: object
    type: object
...
  entity.Entity:
    properties:
      _id:
        type: string
      category:
        $ref: '#/definitions/entitycategory.Category'
        type: object
      component:
        type: string
      component_infos:
        $ref: '#/definitions/entity.Infos'
        type: object
      deletable:
        type: boolean
      depends:
        items:
          type: string
        type: array
      enable_history:
        items:
          type: integer
        type: array
      enabled:
        type: boolean
      idle_since:
        type: integer
      impact:
        items:
          type: string
        type: array
      impact_level:
        type: integer
      infos:
        $ref: '#/definitions/entity.Infos'
        type: object
      measurements:
        type: object
      name:
        type: string
      type:
        type: string
    type: object
  entity.ExportRequest:
    properties:
      category:
        type: string
      fields:
        $ref: '#/definitions/export.Fields'
        type: object
      filter:
        type: string
      search:
        type: string
      separator:
        type: string
    type: object
  entity.ExportResponse:
    properties:
      _id:
        type: string
      status:
        type: integer
    type: object
  entity.Info:
    properties:
      description:
        type: string
      name:
        type: string
      value:
        type: object
    type: object
  entity.Infos:
    additionalProperties:
      $ref: '#/definitions/entity.Info'
    type: object
...
  entitycategory.Category:
    properties:
      _id:
        type: string
      author:
        type: string
      created:
        type: integer
      name:
        type: string
      updated:
        type: integer
    type: object
...
  export.Field:
    properties:
      label:
        type: string
      name:
        type: string
    type: object
  export.Fields:
    items:
      $ref: '#/definitions/export.Field'
    type: array
  http.Response:
    properties:
      body:
        type: string
      close:
        type: boolean
      contentLength:
        type: integer
      header:
        type: Header
      proto:
        type: string
      protoMajor:
        type: integer
      protoMinor:
        type: integer
      request:
        type: Request
      status:
        type: string
      statusCode:
        type: integer
      tls:
        type: string
      trailer:
        type: Header
      transferEncoding:
        items:
          type: string
        type: array
      uncompressed:
        type: boolean
    type: object
...
  pattern.BoolPattern:
    properties:
      set:
        description: Set is a boolean indicating whether the value was set or not.
        type: boolean
      value:
        description: |-
          Value contains the value of the bool. It should only be taken into
          account if Set is true.
        type: boolean
    type: object
  pattern.EntityPattern:
    properties:
      component:
        $ref: '#/definitions/pattern.StringPattern'
        type: object
      componentInfos:
        additionalProperties:
          $ref: '#/definitions/pattern.InfoPattern'
        type: object
      enabled:
        $ref: '#/definitions/pattern.BoolPattern'
        type: object
      id:
        $ref: '#/definitions/pattern.StringPattern'
        type: object
      infos:
        additionalProperties:
          $ref: '#/definitions/pattern.InfoPattern'
        type: object
      name:
        $ref: '#/definitions/pattern.StringPattern'
        type: object
      shouldBeNil:
        type: boolean
      shouldNotBeNil:
        description: |-
          ShouldNotBeNil is a boolean indicating that the entity should not be
          nil, and ShouldBeNil is a boolean indicating that the entity should be
          nil.
          The two booleans are needed to be able to make the difference between
          the case where no entity pattern was defined (in which case the entity
          can be nil or not), the case where a nil entity pattern was defined (in
          which case the entity should be nil), and the case where a non-nil
          entity pattern was defined (in which case the entity should not be nil).
        type: boolean
      type:
        $ref: '#/definitions/pattern.StringPattern'
        type: object
      unexpectedFields:
        additionalProperties: true
        description: |-
          When unmarshalling a BSON document, the fields of this document that are
          not defined in this struct are added to UnexpectedFields.
        type: object
    type: object
  pattern.EntityPatternList:
    properties:
      patterns:
        items:
          $ref: '#/definitions/pattern.EntityPattern'
        type: array
      set:
        description: |-
          Set is a boolean indicating whether the EntityPatternList has been set
          explicitly or not.
        type: boolean
      valid:
        description: |-
          Valid is a boolean indicating whether the event patterns or valid or
          not.
          Valid is also false if the EntityPatternList has not been set.
        type: boolean
    type: object
...
  pattern.InfoPattern:
    properties:
      description:
        $ref: '#/definitions/pattern.StringPattern'
        type: object
      name:
        $ref: '#/definitions/pattern.StringPattern'
        type: object
      shouldNotBeSet:
        type: boolean
      unexpectedFields:
        additionalProperties: true
        description: |-
          When unmarshalling a BSON document, the fields of this document that are
          not defined in this struct are added to UnexpectedFields.
        type: object
      value:
        $ref: '#/definitions/pattern.StringPattern'
        type: object
    type: object
...
  pattern.StringPattern:
    properties:
      equal:
        $ref: '#/definitions/types.OptionalString'
        description: |-
          If Equal is set, the value of a field has to be equal to the value of
          Equal to be matched by the pattern.
        type: object
      regexMatch:
        $ref: '#/definitions/types.OptionalRegexp'
        description: |-
          If RegexMatch is set, the value of a field has to be matched by this
          regular expression to be matched by the pattern.
        type: object
      unexpectedFields:
        additionalProperties: true
        description: |-
          When unmarshalling a BSON document, the fields of this document that are
          not defined in this struct are added to UnexpectedFields.
          UnexpectedFields should always be empty.
        type: object
    type: object
...
  permission.Permission:
    properties:
      _id:
        type: string
      description:
        type: string
      name:
        type: string
      type:
        type: string
    type: object
...
  types.OptionalRegexp:
    properties:
      set:
        description: Set is a boolean indicating whether the value was set or not.
        type: boolean
      value:
        $ref: '#/definitions/utils.RegexExpression'
        description: |-
          Value contains the value of the regular expression. It should only be
          taken into account if Set is true.
        type: object
    type: object
  types.OptionalString:
    properties:
      set:
        description: Set is a boolean indicating whether the value was set or not.
        type: boolean
      value:
        description: |-
          Value contains the value of the string. It should only be taken into
          account if Set is true.
        type: string
    type: object
...
  utils.RegexExpression:
    type: object
...
info:
  contact: {}
  description: This is a Canopsis server.
  license: {}
  title: Canopsis API
  version: "4.0"
paths:
...
  /contextgraph/import:
    put:
      consumes:
      - application/json
      description: Create import task
      operationId: contextgraph-import-create-import
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/contextgraph.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/contextgraph.ImportResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BasicAuth: []
      summary: Create import task
      tags:
      - contextgraph-import
  /contextgraph/import/status/{id}:
    get:
      description: Get import status by id
      operationId: contextgraph-import-get-status-by-id
      parameters:
      - description: import id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/contextgraph.ImportJob'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BasicAuth: []
      summary: Get import status by id
      tags:
      - contextgraph-import
...
  /entities:
    get:
      consumes:
      - application/json
      description: Get paginated list of entities
      operationId: entities-find-all
      parameters:
      - collectionFormat: multi
        in: query
        items:
          type: string
        name: active_columns[]
        type: array
      - in: query
        name: category
        type: string
      - in: query
        name: filter
        type: string
      - in: query
        name: limit
        type: integer
      - in: query
        name: no_events
        type: boolean
      - in: query
        name: page
        type: integer
      - in: query
        name: paginate
        type: boolean
      - in: query
        name: search
        type: string
      - in: query
        name: sort
        type: string
      - in: query
        name: sort_by
        type: string
      - in: query
        name: with_flags
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.PaginatedListResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.Entity'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ValidationErrorResponse'
      security:
      - ApiKeyAuth: []
      - BasicAuth: []
      summary: Find entities
      tags:
      - entities
  /entities/clean:
    post:
      consumes:
      - application/json
      description: Clean disabled entities
      operationId: entities-clean
      parameters:
      - in: query
        name: archive
        required: true
        type: boolean
      - in: query
        name: archive_dependencies
        type: boolean
      produces:
      - application/json
      responses:
        "202": {}
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BasicAuth: []
      summary: Clean disabled entities
      tags:
      - entities
...
  /entity-export:
    post:
      description: Start export entities
      operationId: entities-export-start
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.ExportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ExportResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ValidationErrorResponse'
      security:
      - ApiKeyAuth: []
      - BasicAuth: []
      summary: Start export entities
      tags:
      - entities
  /entity-export/{id}:
    get:
      description: Get status of export entities
      operationId: entities-export-get
      parameters:
      - description: export task id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ExportResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ValidationErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BasicAuth: []
      summary: Get status of export entities
      tags:
      - entities
  /entity-export/{id}/download:
    get:
      description: Get result of export entities
      operationId: entities-export-download
      parameters:
      - description: export task id
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/csv
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ValidationErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BasicAuth: []
      summary: Get result of export entities
      tags:
      - entities
...
  /permissions:
    get:
      consumes:
      - application/json
      description: Get paginated list of permissions
      operationId: permissions-find-all
      parameters:
      - description: current page
        in: query
        name: page
        required: true
        type: integer
      - description: items per page
        in: query
        name: limit
        required: true
        type: integer
      - description: search query
        in: query
        name: search
        type: string
      - description: sort query
        in: query
        name: sort
        type: string
      - description: sort query
        in: query
        name: sort_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.PaginatedListResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/permission.Permission'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ValidationErrorResponse'
      security:
      - ApiKeyAuth: []
      - BasicAuth: []
      summary: Find permissions
      tags:
      - permissions
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-canopsis-authkey
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"